backend/requirements.txt
python-jobspy==1.1.76
flask==3.1.0
mysql-connector-python==9.1.0
apscheduler